Create Exercise4.cs
Exercise #4
1. Create a class called Circle that contains:
a. Two properties: Radius (of the type double) and Color (of the type 
String), with default value of 1.0 and "red", respectively. 
b. Two overloaded constructors - a default constructor with no 
argument, and a constructor which takes a double argument for 
radius. 
c. Two public methods: SetRadius() and SetColor() that sets the 
radius and color, respectively. 
d. One public method: GetArea(), which return the area of this 
instance.
2. Create a class called Color that contains three properties, R, G, B that 
represent the values of red, green, blue channels.
3. Update class Circle, so it has two properties for fill color and border 
color that are represented by Color. (delete previous Property Color and 
update constructors and methods)
4. Create a class Point, that represents a point in 2d space (x,y).
5. Update class Circle, so it has a center point.
6. Create a method in class Circle, IsPointInCircle(Point p), that checks if a 
Point p is inside the circle.
7. Test the above in Main(). Ask user to create a circle, and then enter a 
point. Print if point is inside the circle.